add user router and upadate Other router🍧

+++ controllers/brandController.js
@@ -1,5 +1,5 @@
-const Brand = require("../model/brand");
-const Headphone = require("../model/headphone");
+const Brand = require("../models/brand");
+const Headphone = require("../models/headphone");
 const { validationResult } = require("express-validator");
 
 exports.index = async (req, res, next) => {
@@ -12,7 +12,7 @@ exports.index = async (req, res, next) => {
 
 exports.insert = async (req, res, next) => {
   try {
-    const { name, description } = req.body;
+    const { name, description, shop } = req.body;
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
       const error = new Error("received incorrect information!");
@@ -24,6 +24,7 @@ exports.insert = async (req, res, next) => {
     const brand = new Brand();
     brand.name = name;
     brand.description = description;
+    brand.shop = shop;
 
     // save
     await brand.save();
@@ -36,28 +37,48 @@ exports.insert = async (req, res, next) => {
   }
 };
 
-exports.product = async (req, res, next) => {
-  const headphones = await Headphone.find().populate("brand");
+exports._item = async (req, res, next) => {
+  try {
+    const { id } = req.params;
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      const error = new Error("received incorrect information!");
+      error.statusCode = 422;
+      error.validation = errors.array();
+      throw error;
+    }
+    const headphone = await Headphone.find({
+      brand: id,
+    }).select("name detail ");
 
-  res.status(200).json({
-    data: headphones,
-  });
+    res.status(200).json({
+      data: headphone,
+    });
+  } catch (err) {
+    next(err);
+  }
 };
 
 exports.show = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const headphone = await Headphone.find({
-      brand: id,
-    }).select("name detail ");
-
-    const brand = await Brand.findById(id).populate("headphones");
+    // const headphone = await Headphone.find({
+    //   brand: id,
+    // }).select("name detail ");
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      const error = new Error("received incorrect information!");
+      error.statusCode = 422;
+      error.validation = errors.array();
+      throw error;
+    }
+    const brand = await Brand.findById(id);
 
     const setBrand = {
       id: brand._id,
       name: brand.name,
       description: brand.description,
-      headphones: headphone,
+      // headphones: headphone,
     };
 
     // setBrand.headphone = headphone;
@@ -79,35 +100,52 @@ exports.show = async (req, res, next) => {
 exports.delete = async (req, res, next) => {
   try {
     const { id } = req.params;
-    // delte brand
-    const brandBefore = await Brand.findById(id);
-    const brand = await Brand.findByIdAndDelete(id);
-
-    if (brand.deletedCount === 0) {
-      const error = new Error("There are no brand ID in the information.");
-      error.statusCode = 400;
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      const error = new Error("received incorrect information!");
+      error.statusCode = 422;
+      error.validation = errors.array();
       throw error;
     }
 
-    //delete Headphones in brand
-    const headphones = await Headphone.find({ brand: id });
-    headphones?.map(async (hp) => {
-      const headphone = await Headphone.deleteOne({
-        _id: hp._id,
-      });
+    const brandBefore = await Brand.findById(id);
 
-      if (headphone.deletedCount === 0) {
-        const error = new Error(
-          "There are no headphones ID in the information."
-        );
+    //delete check Headphones in brand
+    const headphones = await Headphone.find({ brand: id });
+    console.log(headphones);
+    if (headphones.length > 0) {
+      const error = new Error(
+        "Can't no delete brand(have some item in brand)🛑"
+      );
+      error.statusCode = 400;
+      throw error;
+    } else {
+      const brand = await Brand.findByIdAndDelete(id);
+
+      if (brand.deletedCount === 0) {
+        const error = new Error("There are no brand ID in the information. 🏢");
         error.statusCode = 400;
         throw error;
       }
-    });
 
-    res.status(200).json({
-      message: `Successfully removed : ${brandBefore.name} ✔`,
-    });
+      res.status(200).json({
+        message: `Successfully removed : ${brandBefore.name} ✔`,
+      });
+    }
+
+    // headphones?.map(async (hp) => {
+    //   const headphone = await Headphone.deleteOne({
+    //     _id: hp._id,
+    //   });
+
+    //   if (headphone.deletedCount === 0) {
+    //     const error = new Error(
+    //       "There are no headphones ID in the information."
+    //     );
+    //     error.statusCode = 400;
+    //     throw error;
+    //   }
+    // });
   } catch (error) {
     next(error);
   }
